delete user

+++ src/components/Dashboard/index.js
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { useNavigate, Link } from "react-router-dom";
 import axios from "axios";
+import Swal from "sweetalert2";
 import Navbar from "./../Navbar";
 import "./style.css";
 
@@ -31,22 +32,47 @@ const Dashboard = () => {
     }
   };
 
-  const deleteUser = async (id) => {
-    try {
-      await axios.delete(`${process.env.REACT_APP_BASE_URL}/users/${id}`, {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      });
-      setUsers(token);
-    } catch (error) {
-      console.log(error);
-    }
+  const deleteUser = (id) => {
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      iconColor: "#D11A2A",
+      showCancelButton: true,
+      confirmButtonText: "Yes, delete it!",
+      confirmButtonColor: "#D11A2A",
+      cancelButtonText: "No, cancel!",
+      reverseButtons: true,
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        await axios.delete(`${process.env.REACT_APP_BASE_URL}/users/${id}`, {
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
+        });
+        getUsers(token);
+        Swal.fire({
+          title: "Deleted!",
+          text: "The user has been deleted",
+          icon: "success",
+          confirmButtonText: "OK",
+          confirmButtonColor: "#457B9D",
+        });
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire({
+          title: "Cancelled",
+          text: "The user is safe :)",
+          icon: "error",
+          confirmButtonText: "OK",
+          confirmButtonColor: "#457B9D",
+        });
+      }
+    });
   };
 
   return (
     <>
-      <Navbar role={role} page='Dashboard'/>
+      <Navbar role={role} page="Dashboard" />
       <div className="wrapper">
         {!token ? (
           <h1>
